Move to a sub directory so lambda definitions can be added
