+++ .gitignore
@@ -106,3 +106,5 @@ venv.bak/
 p3
 
 .vscode
+
+node_modules
