+++ fargate/bin/build-docker-image
